requirements.txt
numpy==1.25.2
pandas==1.5.3
bottleneck==1.3.7
faiss-cpu==1.10.0
requests==2.29.0
beautifulsoup4==4.12.2
urllib3==1.26.16
pdf2image==1.16.3
pytesseract==0.3.13
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
SQLAlchemy==1.4.41
sentence-transformers==2.2.2
pgvector==0.2.3
psycopg2-binary==2.9.9
huggingface_hub==0.8.1
tokenizers==0.10.3
transformers==4.11.3
openai==0.28
python-dotenv==1.0.1
flask-cors==5.0.1
nltk==3.9.1
botbuilder-core==4.16.2
botbuilder-ai==4.16.2
botbuilder-schema==4.16.2
msal==1.31.1